Rename multitrabajos_pages.rb to pages.rb.
